+++ Examples/Felis.Client.Test/Consumers/TestConsumerWithError.cs
@@ -1,5 +1,6 @@
-﻿using Felis.Client.Attributes;
-using Felis.Client.Test.Models;
+﻿using Felis.Client.Test.Models;
+using Felis.Subscriber;
+using Felis.Subscriber.Attributes;
 
 namespace Felis.Client.Test.Consumers;
