+++ Examples/Felis.Client.Test/Felis.Client.Test.csproj
@@ -7,13 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\Felis.Client\Felis.Client.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Microsoft.OpenApi" Version="1.6.14" />
-      <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="6.5.0" />
-      <PackageReference Include="Swashbuckle.AspNetCore.SwaggerUI" Version="6.5.0" />
+      <ProjectReference Include="..\..\Felis\Felis.csproj" />
     </ItemGroup>
 
 </Project>
