+++ Examples/Felis.Router.Test/Felis.Router.Test.csproj
@@ -7,7 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\Felis.Router\Felis.Router.csproj" />
+      <ProjectReference Include="..\..\Felis\Felis.csproj" />
     </ItemGroup>
 
 </Project>
