--- Felis.Client/Attributes/TopicAttribute.cs
@@ -1,11 +1,0 @@
-﻿namespace Felis.Client.Attributes;
-
-public sealed class TopicAttribute : Attribute
-{
-	public string? Value { get; }
-
-	public TopicAttribute(string? value)
-	{
-		Value = value;
-	}
-}
