--- Felis.Client/IConsume.cs
@@ -1,6 +1,0 @@
-﻿namespace Felis.Client;
-
-public interface IConsume<in T> 
-{
-	public void Process(T entity);
-}
