--- Felis.Router/Enums/MessageSendStatus.cs
@@ -1,14 +1,0 @@
-﻿namespace Felis.Router.Enums;
-
-public enum MessageSendStatus
-{
-    MessageReady,
-    NothingToSend,
-    MessageNotFound,
-    MessageNotReady,
-    MessageWithoutTopic,
-    NoConnectionIdAvailableForTopic,
-    ConnectionIdNotFound,
-    MessageSent,
-    MessageNotSent
-}
