--- Felis.Router/Enums/MessageStatus.cs
@@ -1,11 +1,0 @@
-﻿namespace Felis.Router.Enums;
-
-public enum MessageStatus
-{
-    Ready,
-    Sent,
-    Consumed,
-    Processed,
-    Error
-}
-
