--- Felis.Router/Models/Consumer.cs
@@ -1,7 +1,0 @@
-﻿namespace Felis.Router.Models;
-
-public record Consumer(string? Hostname, string? IpAddress, List<string> Topics, bool Unique)
-{
-    public List<string> Topics { get; set; } = Topics;
-}
-
