+++ Felis/Subscriber/Attributes/TopicAttribute.cs
@@ -1,0 +1,11 @@
+﻿namespace Felis.Subscriber.Attributes;
+
+public sealed class TopicAttribute : Attribute
+{
+	public string? Value { get; }
+
+	public TopicAttribute(string? value)
+	{
+		Value = value;
+	}
+}
