+++ Felis/Subscriber/IConsume.cs
@@ -1,0 +1,6 @@
+﻿namespace Felis.Subscriber;
+
+public interface IConsume<in T> 
+{
+	public void Process(T entity);
+}
